+++ rodcom/core/src/main/java/ru/relex/itschool/core/model/Rc_group_member.java
@@ -11,17 +11,17 @@ import org.hibernate.annotations.GenericGenerator;
 @Entity
 public class Rc_group_member {
 
-    private Long group_member_id;
-    private Long group_id;
-    private Long member_id;
-    private Long role_id;
+    private int group_member_id;
+    private int group_id;
+    private int member_id;
+    private int role_id;
     private boolean is_enabled;
     private boolean is_notify;
 
     public Rc_group_member() {
     }
 
-    public Rc_group_member(Long group_id, Long member_id, Long role_id,
+    public Rc_group_member(int group_id, int member_id, int role_id,
                            boolean is_enabled, boolean is_notify) {
         super();
         this.group_id = group_id;
@@ -34,35 +34,35 @@ public class Rc_group_member {
     @Id
     @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="rc_group_member_seq_gen")
     @SequenceGenerator(name="rc_group_member_seq_gen", sequenceName="RC_GROUP_MEMBER_SEQ", allocationSize = 1)
-    public Long getGroup_member_id() {
+    public int getGroup_member_id() {
         return group_member_id;
     }
     
-    public void setGroup_member_id(Long id) {
+    public void setGroup_member_id(int id) {
     	this.group_member_id = id;
     }
 
-    public Long getGroup_Id() {
+    public int getGroup_Id() {
     	return group_id;
     }
 
-    public void setGroup_Id(Long id) {
+    public void setGroup_Id(int id) {
         this.group_id = id;
     }
 
-    public Long getMember_Id() {
+    public int getMember_Id() {
         return member_id;
     }
 
-    public void setMember_Id(Long id) {
+    public void setMember_Id(int id) {
         this.member_id = id;
     }
 
-    public Long getRole_Id() {
+    public int getRole_Id() {
         return role_id;
     }
 
-    public void setRole_Id(Long id) {
+    public void setRole_Id(int id) {
         this.role_id = id;
     }
